fix: simplify template
