--- src/contexts/mod.rs
@@ -1,6 +1,0 @@
-pub mod aggregator;
-pub mod client;
-pub mod combined;
-pub mod eigen_task;
-// TODO: Replace with your context name
-pub mod example_context;
