--- src/jobs/mod.rs
@@ -1,2 +1,0 @@
-pub mod initialize_task;
-pub mod example_task;
